+++ CHANGELOG.md
@@ -15,5 +15,6 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Issue with servers copy/pasted in the servers folder not being linked up to the right folder
 - File not being shown correctly when export is complete
 - Importing MultiMC zips not importing settings [#792]
+- Issue with saveMods throwing NPE
 
 ### Misc